tfg_informatica_fake_news/requirements.txt
python-dotenv>=1.0,<2.0
Telethon>=1.36,<2.0
langdetect>=1.0,<2.0
pytest>=8.0,<9.0
